backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.12.3
bleach==6.1.0
blis==1.0.1
boto3==1.34.145
botocore==1.34.162
cached_path==1.6.4
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.1
cryptography==43.0.1
cycler==0.12.1
cymem==2.0.10
databricks-connect==14.3.2
databricks-sdk==0.31.1
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.9
distlib==0.3.8
distro==1.9.0
docopt==0.6.2
ecdsa==0.19.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.1.0
fastapi==0.110.0
fastjsonschema==2.20.0
filelock==3.13.4
fonttools==4.55.2
frozenlist==1.4.1
fsspec==2024.9.0
gensim==4.3.3
google-api-core==2.23.0
google-auth==2.34.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.6
identify==2.6.0
idna==3.8
iniconfig==2.0.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jose==1.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.35
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langcodes==3.5.0
langsmith==0.1.104
language_data==1.3.0
lmdb==1.5.1
Mako==1.3.5
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.11
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
nlp-topic-builder @ file:///D:/Workspace/1_ML/%40WGU/topic-builder-dev
NlpToolkit-AnnotatedSentence==1.0.38
NlpToolkit-AnnotatedTree==1.0.22
NlpToolkit-Corpus==1.0.25
NlpToolkit-DataStructure==1.0.14
NlpToolkit-DependencyParser==1.0.25
NlpToolkit-Dictionary==1.0.35
NlpToolkit-FrameNet==1.0.8
NlpToolkit-Math==1.0.18
NlpToolkit-MorphologicalAnalysis==1.0.47
NlpToolkit-NamedEntityRecognition==1.0.11
NlpToolkit-ParseTree==1.0.14
NlpToolkit-PropBank==1.0.21
NlpToolkit-SemanticRoleLabeling==1.0.4
NlpToolkit-SentiNet==1.0.18
NlpToolkit-Util==1.0.11
NlpToolkit-WordNet==1.0.24
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
openai==1.52.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pgvector==0.2.5
pickleshare==0.7.5
pillow==10.4.0
pip-check-reqs==2.5.3
pipreqs==0.5.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
preshed==3.0.9
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary==2.9.10
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.18.0
pyparsing==3.2.0
pyspark==3.5.3
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.9.4
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.13.1
sentence-transformers==2.7.0
setuptools==73.0.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.28
srsly==2.4.8
stack-data==0.6.3
starlette==0.36.3
sympy==1.13.2
tenacity==8.5.0
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
torch @ file:///D:/Tools/devTools/Python/pip/torch-2.2.2%2Bcu118-cp312-cp312-win_amd64.whl#sha256=2e32a36a5363c7a9acf058e24442e4033a3fb128de4a90cb0f16baf6681c89f7
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
typer==0.14.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.27.1
virtualenv==20.26.3
wasabi==1.1.3
watchfiles==0.24.0
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websockets==13.0.1
wrapt==1.17.0
yarg==0.1.9
yarl==1.9.4
